Fikser nullpointer i regel 60 (#748)

+++ src/main/kotlin/no/nav/medlemskap/regler/v1/arbeidsforhold/StillingsprosentForStatsborgerskapsgruppe.kt
@@ -3,7 +3,7 @@ package no.nav.medlemskap.regler.v1.arbeidsforhold
 import no.nav.medlemskap.domene.personhistorikk.Personhistorikk
 import no.nav.medlemskap.domene.personhistorikk.Statsborgerskap
 
-fun hentStillingsprosentForStatsborgerskapsgruppe(personhistorikk: Personhistorikk?): Int {
+fun hentStillingsprosentForStatsborgerskapsgruppe(personhistorikk: Personhistorikk?): Double {
     val statsborgerskap = personhistorikk?.statsborgerskap?.filter { it.historisk != true }
     return when {
         erNorskBorger(statsborgerskap) -> Statsborgerskapstype.NORSK_BORGER.stillingsprosent
@@ -15,8 +15,8 @@ fun erNorskBorger(statsborgerskap: List<Statsborgerskap>?): Boolean {
     return statsborgerskap?.any{ it.landkode == "NOR" } == true
 }
 
-enum class Statsborgerskapstype(val stillingsprosent: Int) {
-    NORSK_BORGER(25),
-    EOS_BORGER(100),
-    ANDRE_BORGERE(60)
+enum class Statsborgerskapstype(val stillingsprosent: Double) {
+    NORSK_BORGER(25.0),
+    EOS_BORGER(100.0),
+    ANDRE_BORGERE(60.0)
 }
